+++ product-service/src/main/java/br/edu/atitus/product_service/clients/CurrencyFallback.java
@@ -1,13 +1,14 @@
 package br.edu.atitus.product_service.clients;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Component;
 
 @Component
 public class CurrencyFallback implements CurrencyClient {
 
 	@Override
-	public CurrencyResponse getCurrency(double value, String source, String target) {
+	public CurrencyResponse getCurrency(BigDecimal value, String source, String target) {
 		return null;
 	}
-
 }
